requirements.txt
Flask==3.0.0
Flask-Mail==0.10.0
python-dotenv==1.0.0
reportlab==4.0.7
gunicorn==20.1.0

